requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
pyotp==2.9.0
WTForms==3.1.1
email-validator==2.1.0.post1
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.3
qrcode==7.4.2
flask-bcrypt==1.0.1
